+++ opengl_acquisition_shared_library/opengl_depth_and_color_renderer/src/Library/MotionCaptureLoader/export/bvh_to_csv.c
@@ -507,8 +507,8 @@ int dumpBVHToCSVBody(
                   //Test using :
                   //rm tmp/bvh_test.csv tmp/2d_test.csv && ./BVHTester --from Motions/MotionCapture/01/01_02.bvh  --repeat 0 --selectJoints 17 hip abdomen chest neck head rshoulder relbow rhand lshoulder lelbow lhand rhip rknee rfoot lhip lknee lfoot --csvOrientation right --randomize2D 1000 5000 -35 45 -35 35 135 35 --occlusions --csv tmp test.csv 2d+bvh
                   //value=666; <- highlight the correct
-                  //value=(float) bvh_constrainAngleCentered0((double) value,0);
-                  value=(float) bvh_RemapAngleCentered0((double) value,csvOrientation);
+                  //value=(float) bvh_constrainAngleCentered0((float) value,0);
+                  value=(float) bvh_RemapAngleCentered0((float) value,csvOrientation);
               }
              }
